+++ packages/react-sdk/src/index.ts
@@ -11,6 +11,7 @@ export {
   getArticle,
   getAllTags,
   getArticleBySlugOrId,
+  PCCConvenienceFunctions,
 } from "@pantheon-systems/pcc-sdk-core";
 export * from "@pantheon-systems/pcc-sdk-core/types";