+++ starters/nextjs-starter-ts/pages/experimental/content-only/[...uri].tsx
@@ -1,5 +1,6 @@
 import {
   PantheonProvider,
+  PCCConvenienceFunctions,
   type Article,
 } from "@pantheon-systems/pcc-react-sdk";
 import { NextSeo } from "next-seo";
@@ -8,11 +9,6 @@ import ArticleView from "../../../components/article-view";
 import { PostGrid } from "../../../components/grid";
 import Layout from "../../../components/layout";
 import { Tags } from "../../../components/tags";
-import {
-  getArticleBySlugOrId,
-  getRecommendedArticles,
-} from "../../../lib/Articles";
-import { buildPantheonClientWithGrant } from "../../../lib/PantheonClient";
 import { pantheonAPIOptions } from "../../api/pantheoncloud/[...command]";
 
 interface ArticlePageProps {
@@ -29,7 +25,12 @@ export default function ArticlePage({
   const seoMetadata = getSeoMetadata(article);
 
   return (
-    <PantheonProvider client={buildPantheonClientWithGrant(grant)}>
+    <PantheonProvider
+      client={PCCConvenienceFunctions.buildPantheonClient({
+        isClientSide: true,
+        pccGrant: grant,
+      })}
+    >
       <Layout>
         <NextSeo
           title={seoMetadata.title}
@@ -69,7 +70,7 @@ export async function getServerSideProps({
   const slugOrId = uri[uri.length - 1];
   const grant = pccGrant || cookies["PCC-GRANT"] || null;
 
-  const article = await getArticleBySlugOrId(
+  const article = await PCCConvenienceFunctions.getArticleBySlugOrId(
     slugOrId,
     publishingLevel ? publishingLevel.toString().toUpperCase() : "PRODUCTION",
   );
@@ -101,7 +102,9 @@ export async function getServerSideProps({
     props: {
       article,
       grant,
-      recommendedArticles: await getRecommendedArticles(article.id),
+      recommendedArticles: await PCCConvenienceFunctions.getRecommendedArticles(
+        article.id,
+      ),
     },
   };
 }