+++ starters/nextjs-starter/pages/articles/[...uri].jsx
@@ -1,22 +1,25 @@
-import { PantheonProvider } from "@pantheon-systems/pcc-react-sdk";
+import {
+  PantheonProvider,
+  PCCConvenienceFunctions,
+} from "@pantheon-systems/pcc-react-sdk";
 import { NextSeo } from "next-seo";
 import queryString from "query-string";
 import ArticleView from "../../components/article-view";
 import { PageGrid } from "../../components/grid";
 import Layout from "../../components/layout";
 import { Tags } from "../../components/tags";
-import {
-  getArticleBySlugOrId,
-  getRecommendedArticles,
-} from "../../lib/Articles";
-import { buildPantheonClientWithGrant } from "../../lib/PantheonClient";
 import { pantheonAPIOptions } from "../api/pantheoncloud/[...command]";
 
 export default function ArticlePage({ article, grant, recommendedArticles }) {
   const seoMetadata = getSeoMetadata(article);
 
   return (
-    <PantheonProvider client={buildPantheonClientWithGrant(grant)}>
+    <PantheonProvider
+      client={PCCConvenienceFunctions.buildPantheonClient({
+        isClientSide: true,
+        pccGrant: grant,
+      })}
+    >
       <Layout>
         <NextSeo
           title={seoMetadata.title}
@@ -56,7 +59,7 @@ export async function getServerSideProps({
   const slugOrId = uri[uri.length - 1];
   const grant = pccGrant || cookies["PCC-GRANT"] || null;
 
-  const article = await getArticleBySlugOrId(
+  const article = await PCCConvenienceFunctions.getArticleBySlugOrId(
     slugOrId,
     publishingLevel ? publishingLevel.toString().toUpperCase() : "PRODUCTION",
   );
@@ -88,7 +91,7 @@ export async function getServerSideProps({
     props: {
       article,
       grant,
-      recommendedArticles: await getRecommendedArticles(article.id),
+      recommendedArticles: await PCCConvenienceFunctions.getRecommendedArticles(article.id),
     },
   };
 }