+++ packages/browser/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @pantheon-systems/pcc-browser-sdk
 
+## 3.2.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @pantheon-systems/pcc-sdk-core@3.2.0
+
 ## 3.1.2
 
 ### Patch Changes