+++ packages/react-sample-library/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pantheon-systems/pcc-vue-sdk
 
+## 3.2.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @pantheon-systems/pcc-react-sdk@3.2.0
+  - @pantheon-systems/pcc-sdk-core@3.2.0
+
 ## 3.1.2
 
 ### Patch Changes